+++ src/components/Timeline.vue
@@ -266,7 +266,9 @@ export default defineComponent({
       return this.$route.name === "archive";
     },
     isMonthView(): boolean {
-      return this.$route.name === "albums";
+      return (
+        this.$route.name === "albums" || this.$route.name === "album-share"
+      );
     },
     /** Get view name for dynamic top matter */
     viewName(): string {